Move time state into indicator
